+++ lib/expressions.js
@@ -1,0 +1,18 @@
+module.exports.atxClosedHeaderNoSpaceRe = /(?:^#+\S)|(?:\S#+\s*$)/;
+module.exports.atxClosedHeaderSpaceRe = /(?:^#+\s\s+?\S)|(?:\S\s\s+?#+\s*$)/;
+module.exports.atxHeaderSpaceRe = /^#+\s*\S/;
+module.exports.bareUrlRe = /(?:http|ftp)s?:\/\/[^\s]*/;
+module.exports.dollarCommandRe = /^(\s*)(\$\s)/;
+module.exports.emptyLinkRe = /\[[^\]]*](?:\((?:#?|(?:<>))\))|(?:\[[^\]]*])/;
+module.exports.htmlRe = /<[^>]*>/;
+module.exports.listItemMarkerRe = /^[\s>]*(?:[*+-]|\d+\.)\s+/;
+module.exports.listItemMarkerInterruptsRe = /^[\s>]*(?:[*+-]|1\.)\s+/;
+module.exports.reversedLinkRe = /\([^)]+\)\[[^\]^][^\]]*]/;
+module.exports.spaceAfterBlockQuote = />\s+\S/;
+module.exports.spaceBeforeHeaderRe = /^\s+\S/;
+module.exports.spaceInsideCodeRe = /`(?:(?:\s[^`]*)|(?:[^`]*\s))`/;
+module.exports.spaceInsideEmphasisRe = /(\*\*?|__?)(?:(?:\s.+)|(?:.+\s))\1/;
+module.exports.spaceInsideLinkRe = /\[(?:(?:\s[^\]]*)|(?:[^\]]*\s))](?=\(\S*\))/;
+module.exports.tabRe = /\t+/;
+module.exports.trailingPunctuationRe = /.$/;
+module.exports.trailingSpaceRe = /\s+$/;
